fix: extract wireIdentity to keep wire_services.go under 800-line limit
Move wireIdentity into cmd/meridian/wire_identity.go so that
wire_services.go stays within the 800-line file size limit enforced
by tests/architecture/size_test.go.
